refactor js pull thumbnail related code to thumbnail-factory module use input url first letter for fallback thumbnail add ping to verify input url use nodemon for devtest

+++ index.js
@@ -1,10 +1,10 @@
 const appName = 'Home Tab Bookmark';
 const version = '1.0.0';
 const express = require('express');
-const axios = require('axios');
-const sharp = require('sharp');
-const fs = require('fs');
+const ping = require('ping');
 const url = require('url');
+const fs = require('fs');
+const tf = require('./thumbnail-factory');
 
 // Available ENV
 const screenshotAPI = process.env.SCREENSHOT_API || 'false';
@@ -12,12 +12,11 @@ const openInNewTab = process.env.OPEN_IN_NEW_TAB || 'false';
 const enableBasicAuth = process.env.BASIC_AUTH || 'false';
 const username = process.env.USER_NAME || 'admin';
 const password = process.env.PASSWORD || 'admin';
-
 const port = process.env.PORT || 3000;
-const faviconAPI = 'https://t2.gstatic.com/faviconV2?client=SOCIAL&type=FAVICON&fallback_opts=TYPE,SIZE,URL&size=64&url=';
+
 const dataFile = './data/home-tab-bookmark-data.json';
 const thumbnailDir = './data/thumbnail';
-const sleep = ms => new Promise(res => setTimeout(res, ms));
+
 const authentication = (request, response, next) => {
     if (enableBasicAuth == 'false') {
         next();
@@ -46,82 +45,16 @@ const parseUrlParts = address => {
     return url.parse(address, true);
 };
 
-const getFileNameAndPath = address => {
+const getThumbnailProperties = (address, thumbnailType, screenshotAPI) => {
     const urlParts = parseUrlParts(address);
     const prefix = urlParts.hostname;
     !fs.existsSync(thumbnailDir) && fs.mkdirSync(thumbnailDir);
+    const originUrl = address;
     const originName = `${prefix}-${Date.now()}_origin.png`;
     const originPath = `${thumbnailDir}/${originName}`;
     const destName = `${prefix}-${Date.now()}.png`;
     const destPath = `${thumbnailDir}/${destName}`;
-    return [originName, originPath, destName, destPath];
-};
-
-const useDefaultThumbnail = (destPath, destName, originName) => {
-    try {
-        fs.copyFileSync('./default.thumbnail.png', destPath);
-        return {status: 200, message: 'Error while creating thumbnail, use default thumbnail instead.', filename: destName};
-    }
-    catch (error) {
-        console.log(`error during use default image, ${error.message}`);
-        return {status: 304, message: 'Error while creating thumbnail, please try again later.', filename: originName};
-    }
-};
-
-const downloadImage = async (address, type) => {
-    const [originName, originPath, destName, destPath] = getFileNameAndPath(address);
-    const apiUrl = type == 'icon' ? faviconAPI : ((screenshotAPI && screenshotAPI != '') ? screenshotAPI : faviconAPI);
-    try {
-        return await axios({
-            url: `${apiUrl}${address}`,
-            method: 'GET',
-            responseType: 'stream'
-        })
-        .then(async response => {
-            response.data.pipe(fs.createWriteStream(originPath));
-            const thumbnailWidth = 256;
-            const thumbnailHeight = 192;
-            const originBufferWidth = type == 'icon' ? 64 : 256;
-            const originBufferHeight = type == 'icon' ? 64 : 192;
-            const compositeTop = type == 'icon' ? ((thumbnailHeight - originBufferHeight) / 2 - 10) : 0;
-            const compositeLeft = type == 'icon' ? ((thumbnailWidth - originBufferWidth) / 2) : 0;
-            const baseImage = await sharp({
-                create: {
-                    width: thumbnailWidth,
-                    height: thumbnailHeight,
-                    channels: 4,
-                    background: { r: 0, g: 0, b: 0, alpha: 0 }
-                }
-            })
-            .png()
-            .toBuffer()
-            .catch(error => console.log(`error during creating base image, ${error.message}`));
-            const originImage = await sharp(originPath)
-            .resize(originBufferWidth, originBufferHeight)
-            .png()
-            .toBuffer()
-            .catch(error => console.log(`error during resizing origin image, ${error.message}`));
-            return await sharp(baseImage)
-            .composite([{input: originImage, top: compositeTop, left: compositeLeft}])
-            .toFile(destPath)
-            .then(() => {
-                fs.existsSync(originPath) && fs.unlinkSync(originPath);
-                return {status: 200, filename: destName};
-            })
-            .catch(error => {
-                console.log(`error during composite images, ${error.message}`);
-                return useDefaultThumbnail(destPath, destName, originName);
-            });
-        })
-        .catch(error => {
-            console.log(`error during getting image, ${error.message}`);
-            return useDefaultThumbnail(destPath, destName, originName);
-        });
-    }
-    catch (error) {
-        console.log(`error during getting image, ${error.message}`);
-        return useDefaultThumbnail(destPath, destName, originName);
-    }
+    return {originUrl, thumbnailType, originName, originPath, destName, destPath, screenshotAPI};
 };
 
 const app = express();
@@ -155,14 +88,19 @@ app.get('/process', async (request, response) => {
         const thumbnailType = queryParams.type ? queryParams.type : 'icon';
         if (queryParams.url) {
             try {
-                const result = thumbnailType == 'screen' && typeof puppeteerExtra === 'object' ?
-                await takeScreenshot(queryParams.url) :
-                await downloadImage(queryParams.url, thumbnailType);
-                if (result.status == 200 && queryParams.delete) {
-                    const deleteThumbnailPath = `${thumbnailDir}/${queryParams.delete}`;
-                    fs.existsSync(deleteThumbnailPath) && fs.unlinkSync(deleteThumbnailPath);
+                const pingresult = await ping.promise.probe(queryParams.url.replace(/(https|http):\/\//i, ''));
+                if (pingresult.alive) {
+                    const fileProps = getThumbnailProperties(queryParams.url, thumbnailType, screenshotAPI);
+                    const result = await tf.createThumbnail(fileProps);
+                    if (result.status == 200 && queryParams.delete) {
+                        const deleteThumbnailPath = `${thumbnailDir}/${queryParams.delete}`;
+                        fs.existsSync(deleteThumbnailPath) && fs.unlinkSync(deleteThumbnailPath);
+                    }
+                    response.set('Content-Type', 'text/plain').status(result.status).send(JSON.stringify(result));
                 }
-                response.set('Content-Type', 'text/plain').status(result.status).send(JSON.stringify(result));
+                else {
+                    response.set('Content-Type', 'text/plain').status(404).send(JSON.stringify({status: 404, message: 'Unable reslove the url, please check the url and then try again.'}));
+                }
             }
             catch (error) {
                 // nothing for now
